Scope --config flag to the `buffalo new` command
